task3_moviepicker_sql/requirements.txt
mysql-connector-python==8.0.26
dask==2021.8.1